+++ src/cpp/session/SessionConsoleProcessPersist.cpp
@@ -1,7 +1,7 @@
 /*
  * SessionConsoleProcessPersist.cpp
  *
- * Copyright (C) 2009-17 by RStudio, Inc.
+ * Copyright (C) 2009-18 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -199,7 +199,7 @@ std::string getSavedBuffer(const std::string& handle, int maxLines)
 int getSavedBufferLineCount(const std::string& handle, int maxLines)
 {
    std::string buffer = getSavedBuffer(handle, maxLines);
-   return string_utils::countNewlines(buffer) + 1;
+   return static_cast<int>(string_utils::countNewlines(buffer) + 1);
 }
 
 void appendToOutputBuffer(const std::string& handle, const std::string& buffer)
